feat: deep intelligence optimization - 12 items across P0/P1/P2
P0 Bug Fixes:
1. Fix OpenAI Compatible client missing Tool Calling fields (ToolCalls/ToolCallID/Name/Tools/ToolChoice)
2. Fix Chat() max_tokens type assertion (int -> float64)
3. Fix compressToolOutputs message duplication bug (mark-skip approach)

P1 Intelligence Improvements:
4. RequestOrchestrator - unified intent classification and routing (skill/plan/tool/knowledge/chat)
5. RAG query rewrite pipeline - LLM-based query expansion for better retrieval
6. Knowledge search multi-factor scoring (vectorScore + accessBoost + confidenceBoost + recencyBoost)
7. Auto memory extraction after dialogue completion (integrated MemoryExtractionService)
8. MemoryService Chinese similarity fix (CJK bigram tokenization for Jaccard)

P2 System Architecture:
9. Feedback auto-triggers learning (FeedbackService -> LearningService async pipeline)
10. TokenBudget integration into PromptService (priority-based truncation at 4000 tokens)
11. Prompt layer priority + budget-aware composition
12. LLM call budget check (LLMBudgetManager + periodic detection user limit)

+++ backend/src/services/model_service.go
@@ -551,8 +551,8 @@ func (s *ModelService) Chat(modelID string, messages []llm.Message, options map[
 	if temp, ok := options["temperature"].(float64); ok {
 		req.Temperature = temp
 	}
-	if maxTokens, ok := options["max_tokens"].(int); ok {
-		req.MaxTokens = maxTokens
+	if maxTokens, ok := options["max_tokens"].(float64); ok {
+		req.MaxTokens = int(maxTokens)
 	}
 	if system, ok := options["system"].(string); ok {
 		req.System = system
